Refactored test plugins into tests/plugins, closes #775
