+++ tests/plugins/view_name.py
@@ -1,0 +1,9 @@
+from datasette import hookimpl
+
+
+@hookimpl
+def extra_template_vars(view_name, request):
+    return {
+        "view_name": view_name,
+        "request": request,
+    }
